+++ CHANGELOG.md
@@ -15,6 +15,9 @@
 
 ## `@svizzle/ui` v0.4.0 (next)
 
+- `Link`:
+	- drop `isExternal` prop
+	- add `showIcon` prop
 - `LinkButton`:
 	- add props:
 		- `download`
@@ -22,9 +25,9 @@
 		- `rel`
 		- `target`
 		- `type`
-- `Link`:
-	- drop `isExternal` prop
-	- add `showIcon` prop
+- `LoadingView`:
+	- make sure props are set to their default if `undefined`
+	- add documentation on the site
 
 ## `@svizzle/file` v0.11.0 (next)